Create local room state events in dedicated task
